@@ -1,59 +1,31 @@
 package peaksoft.model;
 
-import javax.persistence.*;
-
-@Table
+import jakarta.persistence.*;
+import lombok.*;
+import lombok.experimental.FieldDefaults;
+@Table(name = "user")
+@Entity
+@Getter
+@Setter
+@ToString
+@AllArgsConstructor
+@NoArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE)
 public class User {
     @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
     private Long id;
-
-    @Column
     private String name;
-
-    @Column
+    @Column(name = "last_name")
     private String lastName;
-
-    @Column
     private Byte age;
 
-    public User() {
-    }
+
 
     public User(String name, String lastName, Byte age) {
         this.name = name;
         this.lastName = lastName;
-        this.age = age;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public Byte getAge() {
-        return age;
-    }
-
-    public void setAge(Byte age) {
         this.age = age;
     }
 
